Group channel creation

+++ src/app/groups.service.ts
@@ -1,6 +1,8 @@
-import { inject, Injectable } from '@angular/core';
+import { createEnvironmentInjector, inject, Injectable } from '@angular/core';
 import { addDoc, collection, collectionData, deleteDoc, doc, Firestore, getDoc, updateDoc } from '@angular/fire/firestore'
 import { Observable } from 'rxjs';
+import { MessagingService } from './messenging.service';
+import { UserService } from './user.service';
 
 @Injectable({
   providedIn: 'root'
@@ -8,6 +10,8 @@ import { Observable } from 'rxjs';
 export class GroupsService {
   private firestore = inject(Firestore);
   private groupsCollection = collection(this.firestore, 'Groups'); // Reference to "Groups" collection 
+  message = inject(MessagingService);
+  user = inject(UserService);
 
   // Get all groups 
   getGroups(): Observable<any[]> {
@@ -31,8 +35,12 @@ export class GroupsService {
     meeting_frequency: string; 
     meeting_times: string; 
     group_members: string[];
+    createdAt: Date;
   }) {
-    return await addDoc(this.groupsCollection, groupData);
+  
+    await addDoc(this.groupsCollection, groupData);
+    this.message.createGroupChannel(this.user.getUid() || '', groupData.group_name, groupData['createdAt'].toISOString().replace(/[^a-zA-Z0-9]/g, ''));
+
   }
 
   // Update a group 
@@ -41,9 +49,57 @@ export class GroupsService {
     return await updateDoc(groupDocRef, updateData);
   }
 
-  // Update a group 
+  // Delete a group 
   async deleteGroup(groupId: string) {
-    const groupDocRef = doc(this.firestore, `Groups/${groupId}`);
-    return await deleteDoc(groupDocRef);
+    
+      const groupData = await this.getGroup(groupId);
+  
+      if (groupData) {
+        const createdAtString = (groupData['createdAt']).toDate().toISOString().replace(/[^a-zA-Z0-9]/g, '');
+        const channelId = (groupData['group_name'] + createdAtString).replace(/\s+/g, '');
+  
+        await this.message.deleteChannel(channelId);
+  
+        const groupDocRef = doc(this.firestore, `Groups/${groupId}`);
+        await deleteDoc(groupDocRef);
+  
+        console.log('Group deleted');
+      } else {
+        console.error('Group not found');
+      
+    }
   }
+
+  // Leave a group
+  async leaveGroup(groupId: string) {
+    const groupData = await this.getGroup(groupId);
+  
+    if (groupData) {
+      const createdAtString = groupData['createdAt'].toDate().toISOString().replace(/[^a-zA-Z0-9]/g, '');
+      const channelId = (groupData['group_name'] + createdAtString).replace(/\s+/g, '');
+      const userId = this.user.getUid();
+  
+     
+        await this.message.removeGroupChannelMember(userId || '', channelId);
+    } else {
+      console.error('Group not found');
+    }
+  }
+
+  async addGroupChannelMember(groupId: string) {
+    const groupData = await this.getGroup(groupId);
+  
+    if (groupData) {
+      const createdAtString = groupData['createdAt'].toDate().toISOString().replace(/[^a-zA-Z0-9]/g, '');
+      const channelId = (groupData['group_name'] + createdAtString).replace(/\s+/g, '');
+      const userId = this.user.getUid();
+  
+     
+        await this.message.addGroupChannelMember(userId || '', channelId);
+    } else {
+      console.error('Group not found');
+    }
+  }
+  
+  
 }
